fix for PR #320 add parent class removed 'try' construction
